Add files via upload
I have created this application using Streamlit, and I have deployed it. This application is a sample of how to analyze student data, manage student data, and how we can store it in the system. It also shows how we can analyze this data according to different criteria. I have provided some background and data for it. 📊🎓